move get_proxy tests from unit test into integration proxy test, and try to use lock to avoid env operation problem.
