+++ tests/client.rs
@@ -4,7 +4,6 @@ extern crate reqwest;
 mod support;
 
 use std::io::Read;
-use std::env;
 
 #[test]
 fn test_response_text() {
@@ -459,36 +458,4 @@ fn test_appended_headers_not_overwritten() {
     assert_eq!(res.status(), reqwest::StatusCode::OK);
     assert_eq!(res.headers().get(reqwest::header::SERVER).unwrap(), &"test");
     assert_eq!(res.headers().get(reqwest::header::CONTENT_LENGTH).unwrap(), &"0");
-}
-
-#[test]
-fn test_system_proxy_is_used() {
-    let server = server! {
-        request: b"\
-            GET http://hyper.rs/prox HTTP/1.1\r\n\
-            user-agent: $USERAGENT\r\n\
-            accept: */*\r\n\
-            accept-encoding: gzip\r\n\
-            host: hyper.rs\r\n\
-            \r\n\
-            ",
-        response: b"\
-            HTTP/1.1 200 OK\r\n\
-            Server: proxied\r\n\
-            Content-Length: 0\r\n\
-            \r\n\
-            "
-    };
-    // set-up http proxy first.
-    env::set_var("http_proxy", format!("http://{}", server.addr()));
-
-    let url = "http://hyper.rs/prox";
-    let res = reqwest::get(url).unwrap();
-
-    assert_eq!(res.url().as_str(), url);
-    assert_eq!(res.status(), reqwest::StatusCode::OK);
-    assert_eq!(res.headers().get(reqwest::header::SERVER).unwrap(), &"proxied");
-
-    // clean system setting
-    env::remove_var("http");
 }
